requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
aiohttp==3.9.3
aiomysql==0.2.0
aiofiles==23.2.1
openai==1.12.0
python-dotenv==1.0.1
websockets==12.0
pandas==1.5.3
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dateutil==2.8.2 